+++ CHANGELOG.md
@@ -7,11 +7,15 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Add tests.
+
 ## [1.10.6] - 2019-03-11
 
 ### Fixed
 
-- Add missing product proptypes.
+- Add missing product propTypes.
 
 ## [1.10.5] - 2019-03-07