+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [1.11.0] - 2019-03-27
+### Added
+- Add `ProductHighlights` component. 
+
 ## [1.10.7] - 2019-03-14
 ### Changed
 - Change basic language files names.